refactor(tenants): extract BillingSection component
